Split up state code
