+++ src/emulator/emulator.cpp
@@ -1,7 +1,7 @@
 #include <faasm/memory.h>
 
 #include <redis/Redis.h>
-#include <state/state.h>
+#include <state/State.h>
 
 /**
  * Local implementations of faasm calls
