+++ src/state/UserState.cpp
@@ -1,0 +1,62 @@
+#include "UserState.h"
+
+#include <util/locks.h>
+
+using namespace util;
+
+namespace state {
+    /**
+     * User state (can have multiple key/ values)
+     */
+
+    UserState::UserState(const std::string &userIn) : user(userIn) {
+
+    }
+
+    UserState::~UserState() {
+        // Delete contents of key value store
+        for (const auto &iter: kvMap) {
+            delete iter.second;
+        }
+
+        kvMap.clear();
+    }
+
+    StateKeyValue *UserState::getValue(const std::string &key, size_t size) {
+        if (kvMap.count(key) == 0) {
+            if (size == 0) {
+                throw std::runtime_error("Must provide a size when getting a value that's not already present");
+            }
+
+            // Lock on editing local state registry
+            FullLock fullLock(kvMapMutex);
+
+            std::string actualKey = user + "_" + key;
+
+            // Double check it still doesn't exist
+            if (kvMap.count(key) == 0) {
+                auto kv = new StateKeyValue(actualKey, size);
+
+                kvMap.emplace(KVPair(key, kv));
+            }
+        }
+
+        return kvMap[key];
+    }
+
+    void UserState::pushAll() {
+        // Iterate through all key-values
+        SharedLock sharedLock(kvMapMutex);
+
+        for (const auto &kv : kvMap) {
+            // Attempt to push partial updates
+            kv.second->pushPartial();
+
+            // Attempt to push partial updates
+            kv.second->pushFull();
+
+            // Attempt to clear (will be ignored if not relevant)
+            kv.second->clear();
+        }
+    }
+}
